Some tidy up work;

+++ period_search_optimization_simd/period_search/memory.c
@@ -1,6 +1,6 @@
 /* allocation and deallocation of memory
 
-   24.10.2005 consulted with Lada Subr 
+   24.10.2005 consulted with Lada Subr
 
    8.11.2006
 */
@@ -23,7 +23,7 @@ double *vector_double(int length)
    }
    return (p_x);
 }
-  
+
 int *vector_int(int length)
 {
    int *p_x;
@@ -40,11 +40,11 @@ double **matrix_double(int rows, int columns)
 {
    double **p_x;
    int i;
-  
+
    p_x = (double **)malloc((rows + 1) * sizeof(double *));
-   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++) 
-      p_x[i] = (double *) malloc((columns + 1) * sizeof(double));
-   if (i < rows) 
+   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++)
+	   p_x[i] = (double*)malloc((static_cast<unsigned long long>(columns) + 1) * sizeof(double));
+   if (i < rows)
    {
       fprintf(stderr,"failure in 'matrix_double()' \n");
       fflush(stderr);
@@ -56,21 +56,21 @@ double **aligned_matrix_double(int rows, int columns)
 {
    double **p_x;
    int i;
-  
+
 #if !defined _WIN32 && !defined __APPLE__
    p_x = (double **)memalign(64,(rows + 1) * sizeof(double *));
-   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++) 
+   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++)
       p_x[i] = (double *) memalign(64,(columns + 1) * sizeof(double));
 #elif defined __APPLE__
    posix_memalign((void **)&p_x, 64,(rows + 1) * sizeof(double *));
-   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++) 
+   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++)
       posix_memalign((void **)&p_x[i], 64,(columns + 1) * sizeof(double));
 #else
    p_x = (double **)_aligned_malloc((rows + 1) * sizeof(double *),64);
-   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++) 
+   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++)
       p_x[i] = (double *) _aligned_malloc((columns + 1) * sizeof(double),64);
 #endif
-   if (i < rows) 
+   if (i < rows)
    {
       fprintf(stderr,"failure in 'matrix_double()' \n");
       fflush(stderr);
@@ -84,8 +84,8 @@ int **matrix_int(int rows, int columns)
    int i;
 
    p_x = (int **) malloc((rows + 1) * sizeof(int *));
-   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++) 
-      p_x[i] = (int *) malloc((columns + 1) * sizeof(int));
+   for (i = 0; (i <= rows) && (!i || p_x[i-1]); i++)
+	   p_x[i] = (int*)malloc((static_cast<unsigned long long>(columns) + 1) * sizeof(int));
    if (i < rows)
    {
       fprintf(stderr,"failure in 'matrix_int()' \n");
@@ -98,20 +98,20 @@ double ***matrix_3_double(int n_1, int n_2, int n_3)
 {
    int i, j;
    double ***p_x;
- 
-   p_x = (double ***) malloc((n_1 + 1) * sizeof(double **));  
-   for (i = 0; i <= n_1; i++)   
+
+   p_x = (double ***) malloc((n_1 + 1) * sizeof(double **));
+   for (i = 0; i <= n_1; i++)
    {
       p_x[i] = (double **) malloc((n_2 + 1) * sizeof(double *));
-      for (j = 0; j <= n_2; j++) 
+      for (j = 0; j <= n_2; j++)
          p_x[i][j] = (double *)malloc((n_3 + 1) * sizeof(double));
-      if (j < n_2) 
+      if (j < n_2)
       {
          fprintf(stderr,"failure in 'matrix_3_double' \n");
          fflush(stderr);
       }
    }
-   if (i < n_1) 
+   if (i < n_1)
    {
       fprintf(stderr,"failure in 'matrix_3_double' \n");
       fflush(stderr);
@@ -119,7 +119,7 @@ double ***matrix_3_double(int n_1, int n_2, int n_3)
 
    return (p_x);
 }
-  
+
 void deallocate_vector(void *p_x)
 {
    free((void *) p_x);
@@ -128,7 +128,7 @@ void deallocate_vector(void *p_x)
 void deallocate_matrix_double(double **p_x, int rows)
 {
    int i;
-    
+
    for (i = 0; i <= rows; i++) free(p_x[i]);
    free(p_x);
 }
@@ -136,7 +136,7 @@ void deallocate_matrix_double(double **p_x, int rows)
 void aligned_deallocate_matrix_double(double **p_x, int rows)
 {
    int i;
-    
+
 #if !defined _WIN32
    for (i = 0; i <= rows; i++) free(p_x[i]);
    free(p_x);
@@ -149,7 +149,7 @@ void aligned_deallocate_matrix_double(double **p_x, int rows)
 void deallocate_matrix_int(int **p_x, int rows)
 {
    int i;
-   
+
    for (i = 0; i <= rows; i++) free(p_x[i]);
    free(p_x);
 }
@@ -158,9 +158,9 @@ void deallocate_matrix_int(int **p_x, int rows)
 /*void deallocate_matrix_3(void ***p_x, int n_1, int n_2)
 {
    int i, j;
-    
+
    for (i = 0; i <= n_1; i++)
-   {  
+   {
       for (j = 1; j <= n_2; j++)
       {
          free(p_x[i][j]);
